Organize dependencies (now main stuff doesn't depend on mixin/asm any more)
Support slim tinytree without param/doc, change tinymetadata index behaviors
