+++ build.gradle
@@ -22,6 +22,12 @@ version = version + "+" + (buildNumber ? "build.$buildNumber" : "local")
 group = 'net.fabricmc'
 archivesBaseName = 'tiny-mappings-parser'
 
+sourceSets {
+	asm
+	mixin
+	legacy // old deprecated stuff, to be removed in 0.4
+}
+
 repositories {
 	mavenCentral()
 	maven {
@@ -32,18 +38,35 @@ repositories {
 
 dependencies {
 	compileOnly 'org.checkerframework:checker-qual:3.3.0'
-	implementation 'org.ow2.asm:asm:8.0'
-	implementation 'org.ow2.asm:asm-analysis:8.0'
-	implementation 'org.ow2.asm:asm-commons:8.0'
-	implementation 'org.ow2.asm:asm-tree:8.0'
-	implementation 'org.ow2.asm:asm-util:8.0'
-	implementation('net.fabricmc:sponge-mixin:0.7.11.17') {
+	// source-set specific stuff aren't exposed later
+	asmImplementation 'org.ow2.asm:asm:8.0'
+	asmImplementation 'org.ow2.asm:asm-analysis:8.0'
+	asmImplementation 'org.ow2.asm:asm-commons:8.0'
+	asmImplementation 'org.ow2.asm:asm-tree:8.0'
+	asmImplementation 'org.ow2.asm:asm-util:8.0'
+	mixinImplementation('net.fabricmc:sponge-mixin:0.7.11.17') {
 		transitive = false
 	}
 
 	testImplementation 'com.google.guava:guava:29.0-jre'
 	testImplementation 'org.junit.jupiter:junit-jupiter-api:5.1.0'
 	testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.1.0'
+}
+
+sourceSets {
+	asm.compileClasspath += main.compileClasspath
+	asm.compileClasspath += main.output
+	mixin.compileClasspath += main.compileClasspath
+	mixin.compileClasspath += main.output
+	legacy.compileClasspath += main.compileClasspath
+	legacy.compileClasspath += asm.compileClasspath
+	legacy.compileClasspath += mixin.compileClasspath
+	test.compileClasspath += asm.output
+	test.compileClasspath += mixin.output
+	test.compileClasspath += legacy.output
+	test.runtimeClasspath += asm.output
+	test.runtimeClasspath += mixin.output
+	test.runtimeClasspath += legacy.output
 }
 
 javadoc {
@@ -56,20 +79,32 @@ javadoc {
 		charSet = 'UTF-8'
 		memberLevel = JavadocMemberLevel.PROTECTED
 		links(
-				'https://asm.ow2.io/javadoc/',
 				'https://checkerframework.org/api/',
-				'https://docs.oracle.com/javase/8/docs/api/'
+				'https://docs.oracle.com/javase/8/docs/api/',
+				'https://asm.ow2.io/javadoc/'
 		)
 		// Disable the crazy super-strict doclint tool in Java 8
 		addStringOption('Xdoclint:none', '-quiet')
 	}
-	source sourceSets.main.allJava.srcDirs
-	classpath = sourceSets.main.compileClasspath
+	source sourceSets.main.allJava.srcDirs, sourceSets.asm.allJava.srcDirs, sourceSets.mixin.allJava.srcDirs
+	classpath += sourceSets.main.compileClasspath
+	classpath += sourceSets.asm.compileClasspath
+	classpath += sourceSets.mixin.compileClasspath
+}
+
+jar {
+	from sourceSets.main.output
+	from sourceSets.asm.output
+	from sourceSets.mixin.output
+	from sourceSets.legacy.output
 }
 
 task sourcesJar(type: Jar, dependsOn: classes) {
 	archiveClassifier.set 'sources'
 	from sourceSets.main.allSource
+	from sourceSets.asm.allSource
+	from sourceSets.mixin.allSource
+	from sourceSets.legacy.allSource
 }
 
 task javadocsJar(type: Jar, dependsOn: javadoc) {
@@ -78,6 +113,7 @@ task javadocsJar(type: Jar, dependsOn: javadoc) {
 }
 
 artifacts {
+	archives jar
 	archives sourcesJar
 	archives javadocsJar
 }
@@ -91,7 +127,6 @@ publishing {
 		maven(MavenPublication) {
 			from components.java
 
-			artifact jar
 			artifact sourcesJar
 			artifact javadocsJar
 		}
