+++ src/test/java/net/fabricmc/mappings/tests/MappingsInterfaceTest.java
@@ -16,17 +16,21 @@
 
 package net.fabricmc.mappings.tests;
 
-import net.fabricmc.mappings.EntryTriple;
-import org.junit.jupiter.api.Assertions;
+import net.fabricmc.mappings.Mappings;
+import net.fabricmc.mappings.model.V2MappingsProvider;
 import org.junit.jupiter.api.Test;
 
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
+import java.nio.charset.StandardCharsets;
+
 @Deprecated
-class EntryTripleTests {
-    @Test
-    void equalsTest() {
-        Assertions.assertEquals(
-                new EntryTriple("a", "b", "c"),
-                new EntryTriple("a", "b", "c")
-        );
+public class MappingsInterfaceTest {
+	@Test
+	public void testParsingToOldInterface() throws Exception {
+		try (BufferedReader reader = new BufferedReader(new InputStreamReader(MappingsInterfaceTest.class.getResourceAsStream("/bigboi_mappings.tinyv2"), StandardCharsets.UTF_8))) {
+            Mappings mappings = V2MappingsProvider.readTinyMappings(reader);
+            int x = 2;
+        }
     }
 }
